перенесла типы
