+++ src/services/types/types.ts
@@ -46,3 +46,218 @@ export type TCommentRaw = {
   text: string;
   to: TUserInfoRaw;
 }
+// типы данных пришедших с сервера
+export type BaseFiedsRaw = {
+  _id: string,
+  createdAt: number,
+  updatedAt: number,
+}
+export type UserAccountRaw = {
+  email: string,
+  cohort: string,
+}
+export type GeocodeRaw = Array<number>
+export type ShortProfileRaw = {
+  name: string,
+  photo: string,
+  city: {
+    name: string,
+    geocode: GeocodeRaw,
+  }
+}
+export type ProfileRaw = {
+  name: string,
+  photo: string,
+  city: {
+    name: string,
+    geocode: GeocodeRaw,
+  }
+  birthday: string,
+  quote: string,
+  telegram: string,
+  github: string,
+  template: string,
+}
+export type InfoBlocksRaw = {
+  hobby: {
+    text?: string,
+    image?: string,
+    reactions?: number
+  }
+  status: {
+    text?: string,
+    image?: string,
+    reactions?: number
+  }
+  job: {
+    text?: string,
+    image?: string,
+    reactions?: number
+  }
+  edu: {
+    text?: string,
+    image?: string,
+    reactions?: number
+  }
+}
+export type InfoItemsRaw ={
+  hobby: {
+    text?: string,
+    image?: string | null,
+  }
+  status: {
+    text?: string,
+    image?: string | null,
+  }
+  job: {
+    text?: string,
+    image?: string | null,
+  }
+  edu: {
+    text?: string,
+    image?: string | null,
+  }
+}
+export type UserWithProfileRaw = {
+  email: string,
+  cohort: string,
+  profile:{
+    name: string,
+    photo: string,
+    city: {
+      name: string,
+      geocode: GeocodeRaw,
+    }
+    birthday: string,
+    quote: string,
+    telegram: string,
+    github: string,
+    template: string,
+  }
+  info: {
+    hobby: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+    status: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+    job: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+    edu: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+  }
+}
+export type Reactions = Array<CommentRaw | LikeRaw>
+export type UserRaw = {
+  email: string,
+  cohort: string,
+  profile:{
+    name: string,
+    photo: string,
+    city: {
+      name: string,
+      geocode: GeocodeRaw,
+    }
+    birthday: string,
+    quote: string,
+    telegram: string,
+    github: string,
+    template: string,
+  }
+  info: {
+    hobby: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+    status: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+    job: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+    edu: {
+      text?: string,
+      image?: string,
+      reactions?: number
+    }
+  }
+  reactions: Reactions
+}
+export type UserRefRaw = {
+  _id: string,
+  name: string,
+  email: string
+}
+export type TargetRaw = 'hobby'|'status'|'job'|'edu';
+export type ReactionRaw = {
+  _id: string,
+  from:{
+    _id: string,
+    name: string,
+    email: string
+  },
+  target: TargetRaw,
+}
+export type CommentRaw = {
+  _id: string,
+  from:{
+    _id: string,
+    name: string,
+    email: string
+  },
+  target: TargetRaw,
+  text: string,
+}
+export type LikeRaw = {
+  _id: string,
+  from:{
+    _id: string,
+    name: string,
+    email: string
+  },
+  target: TargetRaw,
+  emotion: string
+}
+export type InfoItemRaw = {
+  text?: string,
+  image?: string,
+}
+export type InfoBlockRaw = {
+  text?: string,
+  image?: string,
+  reactions?: number
+}
+///
+export type TGetUsersRaw = {
+  total: number,
+  items: Array<BaseFiedsRaw & UserAccountRaw & {name: string}>
+}
+export type TGetCommentsRaw = {
+  total: number,
+  items: Array<CommentRaw & {to: UserRefRaw}>,
+}
+
+export type TUserReactionsRaw = {
+  total: number,
+  items: Array<CommentRaw & LikeRaw>
+}
+export type TGetProfilesRaw = {
+  total: number,
+  items: Array<BaseFiedsRaw & UserAccountRaw & {profile: ShortProfileRaw}>
+}
+
