+++ src/utils/api.ts
@@ -1,3 +1,5 @@
+import { BaseFiedsRaw, InfoItemsRaw, ProfileRaw, TGetCommentsRaw, TGetProfilesRaw, TGetUsersRaw, TUserReactionsRaw, UserAccountRaw, UserWithProfileRaw } from "../services/types/types";
+
 //тестовый токен
 export const token = 'wertyu45678cfgh567'
 //проверка ответа от сервера
@@ -20,220 +22,6 @@ const request = <T>(url: string, options: RequestInit, readBody: boolean = true)
   return fetch(url, options).then(res => checkResponse<T>(res, readBody))
 }
 //TODO: надо разобраться с типами
-type BaseFiedsRaw = {
-  _id: string,
-  createdAt: number,
-  updatedAt: number,
-}
-type UserAccountRaw = {
-  email: string,
-  cohort: string,
-}
-type GeocodeRaw = Array<number>
-type ShortProfileRaw = {
-  name: string,
-  photo: string,
-  city: {
-    name: string,
-    geocode: GeocodeRaw,
-  }
-}
-type ProfileRaw = {
-  name: string,
-  photo: string,
-  city: {
-    name: string,
-    geocode: GeocodeRaw,
-  }
-  birthday: string,
-  quote: string,
-  telegram: string,
-  github: string,
-  template: string,
-}
-type InfoBlocksRaw = {
-  hobby: {
-    text?: string,
-    image?: string,
-    reactions?: number
-  }
-  status: {
-    text?: string,
-    image?: string,
-    reactions?: number
-  }
-  job: {
-    text?: string,
-    image?: string,
-    reactions?: number
-  }
-  edu: {
-    text?: string,
-    image?: string,
-    reactions?: number
-  }
-}
-type InfoItemsRaw ={
-  hobby: {
-    text?: string,
-    image?: string | null,
-  }
-  status: {
-    text?: string,
-    image?: string | null,
-  }
-  job: {
-    text?: string,
-    image?: string | null,
-  }
-  edu: {
-    text?: string,
-    image?: string | null,
-  }
-}
-type UserWithProfileRaw = {
-  email: string,
-  cohort: string,
-  profile:{
-    name: string,
-    photo: string,
-    city: {
-      name: string,
-      geocode: GeocodeRaw,
-    }
-    birthday: string,
-    quote: string,
-    telegram: string,
-    github: string,
-    template: string,
-  }
-  info: {
-    hobby: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-    status: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-    job: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-    edu: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-  }
-}
-type Reactions = Array<CommentRaw | LikeRaw>
-type UserRaw = {
-  email: string,
-  cohort: string,
-  profile:{
-    name: string,
-    photo: string,
-    city: {
-      name: string,
-      geocode: GeocodeRaw,
-    }
-    birthday: string,
-    quote: string,
-    telegram: string,
-    github: string,
-    template: string,
-  }
-  info: {
-    hobby: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-    status: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-    job: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-    edu: {
-      text?: string,
-      image?: string,
-      reactions?: number
-    }
-  }
-  reactions: Reactions
-}
-type UserRefRaw = {
-  _id: string,
-  name: string,
-  email: string
-}
-export type TargetRaw = 'hobby'|'status'|'job'|'edu';
-type ReactionRaw = {
-  _id: string,
-  from:{
-    _id: string,
-    name: string,
-    email: string
-  },
-  target: TargetRaw,
-}
-type CommentRaw = {
-  _id: string,
-  from:{
-    _id: string,
-    name: string,
-    email: string
-  },
-  target: TargetRaw,
-  text: string,
-}
-type LikeRaw = {
-  _id: string,
-  from:{
-    _id: string,
-    name: string,
-    email: string
-  },
-  target: TargetRaw,
-  emotion: string
-}
-type InfoItemRaw = {
-  text?: string,
-  image?: string,
-}
-type InfoBlockRaw = {
-  text?: string,
-  image?: string,
-  reactions?: number
-}
-///
-type TGetUsersRaw = {
-  total: number,
-  items: Array<BaseFiedsRaw & UserAccountRaw & {name: string}>
-}
-type TGetCommentsRaw = {
-  total: number,
-  items: Array<CommentRaw & {to: UserRefRaw}>,
-}
-
-type TUserReactionsRaw = {
-  total: number,
-  items: Array<CommentRaw & LikeRaw>
-}
-type TGetProfilesRaw = {
-  total: number,
-  items: Array<BaseFiedsRaw & UserAccountRaw & {profile: ShortProfileRaw}>
-}
-
 
 
 //запрос всех пользователей
